refactor: separate tests_text.py module
